Refactor current API for component types
